+++ jdaviz/components/tooltip.vue
@@ -66,7 +66,7 @@ const tooltips = {
   'viewer-data-radio': 'Switch visibility to layers associated with this data entry',
   'viewer-data-enable': 'Load data entry into this viewer',
   'viewer-data-disable': 'Disable data within this viewer (will be hidden and unavailable from plugins until re-enabled)',
-  'viewer-data-delete': 'Remove data entry across entire app',
+  'viewer-data-delete': 'Remove data entry across entire app (might affect existing subsets)',
 
   'table-prev': 'Select previous row in table',
   'table-next': 'Select next row in table',